перевела код в TS
