--- src/App.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import HeaderApp from './Layouts/Header';
-import ContentApp from './Layouts/Content';
-
-const App = () => {
-  return (
-    <>
-      <HeaderApp />
-      <ContentApp />
-    </>
-  );
-};
-
-export default App;
