+++ src/App.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import HeaderApp from './Layouts/Header.tsx';
+import ContentApp from './Layouts/Content.tsx';
+
+const App: React.FC = () => {
+  return (
+    <>
+      <HeaderApp />
+      <ContentApp />
+    </>
+  );
+};
+
+export default App;
